requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==3.0.1
Django==5.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==25.0
parameterized==0.9.0
pillow==10.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==70.2.0
sqlparse==0.5.1
uritemplate==4.2.0
urllib3==2.2.2
